+++ logprep/ng/processor/amides/processor.py
@@ -125,7 +125,7 @@ class Amides(FieldManager):
         and the rule attributor.
 
         .. security-best-practice::
-           :title: Processor - Amides Model
+           :title: |PROCESSOR| - Amides Model
 
            Ensure that you only use models from trusted sources, as it can be used to inject python
            code into the runtime.