+++ logprep/processor/domain_resolver/processor.py
@@ -80,7 +80,7 @@ class DomainResolver(Processor):
         """Timeout for resolving of domains.
 
         .. security-best-practice::
-           :title: Processor - Domain Resolver Timeout
+           :title: |PROCESSOR| - Timeout
 
            Ensure to set this to a reasonable value to avoid DOS attacks by malicious domains in
            your logs. The default is set to 0.5 seconds.
@@ -91,7 +91,7 @@ class DomainResolver(Processor):
         Logprep does therefore clear the cache.
 
         .. security-best-practice::
-           :title: Processor - Domain Resolver Max Cached Domains
+           :title: |PROCESSOR| - Max Cached Domains
 
            Ensure to set this to a reasonable value to avoid excessive memory usage
            and OOM situations by the domain resolver cache.