+++ logprep/processor/generic_adder/rule.py
@@ -128,14 +128,14 @@ class GenericAdderRule(FieldManagerRule):
         All of those files must exist. For string format see :ref:`getters`
 
         .. security-best-practice::
-           :title: Processor - Generic Adder Add From File Memory Consumption
+           :title: |PROCESSOR| - Add From File Memory Consumption
 
            Be aware that all values of the remote file were loaded into memory. Consider to avoid
            dynamic increasing lists without setting limits for Memory consumption. Additionally
            avoid loading large files all at once to avoid exceeding http body limits.
 
         .. security-best-practice::
-           :title: Processor - Generic Adder Authenticity and Integrity
+           :title: |PROCESSOR| - Authenticity and Integrity
 
            Consider to use TLS protocol with authentication via mTLS or Oauth to ensure
            authenticity and integrity of the loaded values.