+++ logprep/processor/generic_resolver/processor.py
@@ -59,10 +59,10 @@ class GenericResolver(FieldManager):
         The cache can be disabled by setting this option to :code:`0`.
 
         .. security-best-practice::
-           :title: Processor - Generic Resolver Max Cached Entries
+           :title: |PROCESSOR| - Max Cached Entries
 
            Ensure to set this to a reasonable value to avoid excessive memory usage
-           and OOM situations caused by the generic resolver cache.
+           and OOM situations caused by the |PROCESSOR_NAME| cache.
 
         """
         cache_metrics_interval: int = field(validator=validators.instance_of(int), default=1)